moved from types to existing options file
